game-flow: merge M_HandlePlayLevel to trx
