+++ src/tr1/game/stats/common.h
@@ -4,10 +4,6 @@
 #include <libtrx/game/stats.h>
 
 void Stats_ObserveRoomsLoad(void);
-void Stats_CalculateStats(void);
-int32_t Stats_GetMaxPickups(void);
-int32_t Stats_GetMaxKillables(void);
-int32_t Stats_GetMaxSecrets(void);
 void Stats_ComputeFinal(GF_LEVEL_TYPE level_type, FINAL_STATS *stats);
 bool Stats_CheckAllSecretsCollected(GF_LEVEL_TYPE level_type);
