+++ src/tr2/game/game_flow/sequencer_events.c
@@ -11,7 +11,6 @@
 #include <libtrx/game/option/passport.h>
 #include <libtrx/game/output.h>
 
-static DECLARE_GF_EVENT_HANDLER(M_HandlePlayLevel);
 static DECLARE_GF_EVENT_HANDLER(M_HandlePlayMusic);
 static DECLARE_GF_EVENT_HANDLER(M_HandleLevelComplete);
 static DECLARE_GF_EVENT_HANDLER(M_HandleEnableSunset);
@@ -20,7 +19,6 @@ static DECLARE_GF_EVENT_HANDLER(M_HandleSetStartAnim);
 
 static DECLARE_GF_EVENT_HANDLER((*m_EventHandlers[GFS_NUMBER_OF])) = {
     // clang-format off
-    [GFS_LOOP_GAME]        = M_HandlePlayLevel,
     [GFS_PLAY_MUSIC]       = M_HandlePlayMusic,
     [GFS_LEVEL_COMPLETE]   = M_HandleLevelComplete,
     [GFS_ENABLE_SUNSET]    = M_HandleEnableSunset,
@@ -28,83 +26,6 @@ static DECLARE_GF_EVENT_HANDLER((*m_EventHandlers[GFS_NUMBER_OF])) = {
     [GFS_SET_START_ANIM]   = M_HandleSetStartAnim,
     // clang-format on
 };
-
-static DECLARE_GF_EVENT_HANDLER(M_HandlePlayLevel)
-{
-    GF_COMMAND gf_cmd = { .action = GF_NOOP };
-
-    if (seq_ctx == GFSC_STORY) {
-        const int32_t savegame_level_num = (int32_t)(intptr_t)seq_ctx_arg;
-        if (savegame_level_num == level->num) {
-            return (GF_COMMAND) { .action = GF_EXIT_TO_TITLE };
-        } else {
-            return (GF_COMMAND) { .action = GF_NOOP };
-        }
-    }
-
-    if (Lara_GetItem() != nullptr) {
-        Lara_Initialise(level);
-    }
-
-    if (level->music_track != MX_INACTIVE) {
-        Music_Stop();
-    }
-
-    Lua_FireEvent(LUA_EVENT_LEVEL_LOAD, level->num);
-
-    switch (seq_ctx) {
-    case GFSC_SAVED: {
-        const int16_t slot_num = Savegame_GetBoundSlot();
-        if (!Savegame_Load(slot_num)) {
-            LOG_ERROR("Failed to load save file!");
-            Game_SetCurrentLevel(nullptr);
-            GF_SetCurrentLevel(nullptr);
-            return (GF_COMMAND) { .action = GF_EXIT_TO_TITLE };
-        }
-        break;
-    }
-
-    default:
-        if (level->type == GFL_NORMAL || level->type == GFL_BONUS) {
-            Savegame_SetInitialVersion(SAVEGAME_CURRENT_VERSION);
-            GF_InventoryModifier_Scan(Game_GetCurrentLevel());
-            GF_InventoryModifier_Apply(Game_GetCurrentLevel(), GF_INV_REGULAR);
-        }
-        break;
-    }
-
-    if (level->type == GFL_NORMAL || level->type == GFL_BONUS) {
-        Stats_CalculateStats();
-        RESUME_INFO *const resume = Savegame_GetCurrentInfo(level);
-        if (resume != nullptr) {
-            resume->stats.max_secret_count = Stats_GetMaxSecrets();
-            resume->stats.all_secrets_mask = Stats_GetMaxSecretFlags();
-        }
-    }
-
-    Lua_FireEvent(LUA_EVENT_LEVEL_START, level->num);
-
-    g_Passport.ask_for_save = g_Config.gameplay.enable_save_crystals
-        && seq_ctx == GFSC_NORMAL
-        && GF_GetLevelTableType(level->type) == GFLT_MAIN
-        && level != GF_GetFirstLevel() && level != GF_GetGymLevel();
-
-    ASSERT(GF_GetCurrentLevel() == level);
-    if (level->type == GFL_DEMO) {
-        gf_cmd = GF_RunDemo(level->num);
-    } else if (level->type == GFL_CUTSCENE) {
-        gf_cmd = GF_RunCutscene(level->num);
-    } else {
-        if (seq_ctx != GFSC_SAVED && level != GF_GetFirstLevel()) {
-            Lara_RevertToPistolsIfNeeded();
-        }
-        gf_cmd = GF_RunGame(level, seq_ctx);
-    }
-    if (gf_cmd.action == GF_LEVEL_COMPLETE) {
-        gf_cmd.action = GF_NOOP;
-    }
-    return gf_cmd;
-}
 
 static DECLARE_GF_EVENT_HANDLER(M_HandlePlayMusic)
 {
